add GeneralObserver to Twitter - Instagram - Pinterest
