+++ @pauliescanlon/gatsby-mdx-embed/src/components/Instagram/Instagram.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handleInstagrmLoad } from './utils'
 export interface IInstagramProps {
   /** Instagram id */
   instagramId: string
@@ -8,15 +9,16 @@ export interface IInstagramProps {
 export const Instagram: FunctionComponent<IInstagramProps> = ({
   instagramId
 }: IInstagramProps) => (
-  <blockquote
-    className="instagram-media instagram-mdx-embed"
-    data-instgrm-version="12"
-    // data-instgrm-captioned
-  >
-    <a href={`https://instagram.com/p/${instagramId}`}>
-      {typeof window !== 'undefined' && !(window as any).instgrm
-        ? 'Loading'
-        : ''}
-    </a>
-  </blockquote>
+  <GeneralObserver onEnter={() => handleInstagrmLoad()}>
+    <blockquote
+      className="instagram-media instagram-mdx-embed"
+      data-instgrm-version="12"
+    >
+      <a href={`https://instagram.com/p/${instagramId}`}>
+        {typeof window !== 'undefined' && !(window as any).instgrm
+          ? 'Loading'
+          : ''}
+      </a>
+    </blockquote>
+  </GeneralObserver>
 )
