+++ @pauliescanlon/gatsby-mdx-embed/src/components/Pinterest/Pin.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handlePinterestBuild } from './utils'
 export interface IPinProps {
   /** Pinterest id */
   pinId: string
@@ -11,10 +12,12 @@ export const Pin: FunctionComponent<IPinProps> = ({
   pinId,
   size = 'small'
 }: IPinProps) => (
-  <a
-    className="pinterest-pin pinterest-pin-mdx-embed"
-    data-pin-do="embedPin"
-    data-pin-width={size}
-    href={`https://www.pinterest.com/pin/${pinId}`}
-  />
+  <GeneralObserver onEnter={() => handlePinterestBuild()}>
+    <a
+      className="pinterest-pin pinterest-pin-mdx-embed"
+      data-pin-do="embedPin"
+      data-pin-width={size}
+      href={`https://www.pinterest.com/pin/${pinId}`}
+    />
+  </GeneralObserver>
 )
