+++ @pauliescanlon/gatsby-mdx-embed/src/components/Pinterest/PinterestFollowButton.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handlePinterestBuild } from './utils'
 export interface IPinterestFollowButtonProps {
   /** Pinterest username */
   username: string
@@ -8,11 +9,13 @@ export interface IPinterestFollowButtonProps {
 export const PinterestFollowButton: FunctionComponent<IPinterestFollowButtonProps> = ({
   username
 }: IPinterestFollowButtonProps) => (
-  <a
-    className="pinterest-follow-button pinterest-follow-button-mdx-embed"
-    data-pin-do="buttonFollow"
-    href={`https://www.pinterest.com/${username}/`}
-  >
-    {`Follow @${username}`}
-  </a>
+  <GeneralObserver onEnter={() => handlePinterestBuild()}>
+    <a
+      className="pinterest-follow-button pinterest-follow-button-mdx-embed"
+      data-pin-do="buttonFollow"
+      href={`https://www.pinterest.com/${username}/`}
+    >
+      {`Follow @${username}`}
+    </a>
+  </GeneralObserver>
 )
