+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/Tweet.tsx
@@ -1,4 +1,6 @@
 import React, { FunctionComponent } from 'react'
+import { GeneralObserver } from '../GeneralObserver'
+import { handleTwttrLoad } from './utils'
 
 export interface ITweetProps {
   /** Tweet link */
@@ -14,13 +16,19 @@ export const Tweet: FunctionComponent<ITweetProps> = ({
   theme = 'light',
   align = 'left'
 }: ITweetProps) => (
-  <div className="twitter-tweet-mdx-embed" style={{ overflow: 'auto' }}>
-    <blockquote className="twitter-tweet" data-theme={theme} data-align={align}>
-      <a href={`https://twitter.com/${tweetLink}?ref_src=twsrc%5Etfw`}>
-        {typeof window !== 'undefined' && !(window as any).twttr
-          ? 'Loading'
-          : ''}
-      </a>
-    </blockquote>
-  </div>
+  <GeneralObserver onEnter={() => handleTwttrLoad()}>
+    <div className="twitter-tweet-mdx-embed" style={{ overflow: 'auto' }}>
+      <blockquote
+        className="twitter-tweet"
+        data-theme={theme}
+        data-align={align}
+      >
+        <a href={`https://twitter.com/${tweetLink}?ref_src=twsrc%5Etfw`}>
+          {typeof window !== 'undefined' && !(window as any).twttr
+            ? 'Loading'
+            : ''}
+        </a>
+      </blockquote>
+    </div>
+  </GeneralObserver>
 )
