+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterFollowButton.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handleTwttrLoad } from './utils'
 export interface ITwitterFollowButtonProps {
   /** Twitter username */
   username: string
@@ -17,11 +18,13 @@ export const TwitterFollowButton: FunctionComponent<ITwitterFollowButtonProps> =
   showUsername = true,
   size = 'small'
 }: ITwitterFollowButtonProps) => (
-  <a
-    href={`https://twitter.com/${username}?ref_src=twsrc%5Etfw`}
-    className="twitter-follow-button twitter-follow-button-mdx-embed"
-    data-show-count={showFollowers}
-    data-show-screen-name={showUsername}
-    data-size={size}
-  >{`Follow @${username}`}</a>
+  <GeneralObserver onEnter={() => handleTwttrLoad()}>
+    <a
+      href={`https://twitter.com/${username}?ref_src=twsrc%5Etfw`}
+      className="twitter-follow-button twitter-follow-button-mdx-embed"
+      data-show-count={showFollowers}
+      data-show-screen-name={showUsername}
+      data-size={size}
+    >{`Follow @${username}`}</a>
+  </GeneralObserver>
 )
