+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterHashtagButton.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handleTwttrLoad } from './utils'
 export interface ITwitterHashtagButtonProps {
   /** Twitter hashtag */
   hashtag: string
@@ -11,9 +12,11 @@ export const TwitterHashtagButton: FunctionComponent<ITwitterHashtagButtonProps>
   hashtag,
   size = 'small'
 }: ITwitterHashtagButtonProps) => (
-  <a
-    href={`https://twitter.com/intent/tweet?button_hashtag=${hashtag}&ref_src=twsrc%5Etfw`}
-    className="twitter-hashtag-button twitter-hashtag-button-mdx-embed"
-    data-size={size}
-  >{`Tweet #${hashtag}`}</a>
+  <GeneralObserver onEnter={() => handleTwttrLoad()}>
+    <a
+      href={`https://twitter.com/intent/tweet?button_hashtag=${hashtag}&ref_src=twsrc%5Etfw`}
+      className="twitter-hashtag-button twitter-hashtag-button-mdx-embed"
+      data-size={size}
+    >{`Tweet #${hashtag}`}</a>
+  </GeneralObserver>
 )
