+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterList.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handleTwttrLoad } from './utils'
 export interface ITwitterListProps {
   /** Twitter username */
   username: string
@@ -20,15 +21,17 @@ export const TwitterList: FunctionComponent<ITwitterListProps> = ({
   width = '498px',
   height = null
 }: ITwitterListProps) => (
-  <div style={{ overflow: 'auto' }}>
-    <a
-      className="twitter-timeline twitter-timeline-mdx-embed"
-      data-theme={theme}
-      data-width={width}
-      data-height={height}
-      href={`https://twitter.com/${username}/lists/${listName}?ref_src=twsrc%5Etfw`}
-    >
-      {`A Twitter List by @${username}`}
-    </a>
-  </div>
+  <GeneralObserver onEnter={() => handleTwttrLoad()}>
+    <div style={{ overflow: 'auto' }}>
+      <a
+        className="twitter-timeline twitter-timeline-mdx-embed"
+        data-theme={theme}
+        data-width={width}
+        data-height={height}
+        href={`https://twitter.com/${username}/lists/${listName}?ref_src=twsrc%5Etfw`}
+      >
+        {`A Twitter List by @${username}`}
+      </a>
+    </div>
+  </GeneralObserver>
 )
