+++ @pauliescanlon/gatsby-mdx-embed/src/components/Twitter/TwitterMentionButton.tsx
@@ -1,5 +1,6 @@
 import React, { FunctionComponent } from 'react'
-
+import { GeneralObserver } from '../GeneralObserver'
+import { handleTwttrLoad } from './utils'
 export interface ITwitterMentionButtonProps {
   /** Twitter username */
   username: string
@@ -11,9 +12,11 @@ export const TwitterMentionButton: FunctionComponent<ITwitterMentionButtonProps>
   username,
   size = 'small'
 }: ITwitterMentionButtonProps) => (
-  <a
-    href={`https://twitter.com/intent/tweet?screen_name=${username}&ref_src=twsrc%5Etfw`}
-    className="twitter-mention-button twitter-mention-button-mdx-embed"
-    data-size={size}
-  >{`Tweet to @${username}`}</a>
+  <GeneralObserver onEnter={() => handleTwttrLoad()}>
+    <a
+      href={`https://twitter.com/intent/tweet?screen_name=${username}&ref_src=twsrc%5Etfw`}
+      className="twitter-mention-button twitter-mention-button-mdx-embed"
+      data-size={size}
+    >{`Tweet to @${username}`}</a>
+  </GeneralObserver>
 )
