+++ starter/src/pages/index.mdx
@@ -3,3 +3,6 @@
 #### Look Ma, no imports required
 
 <Gist gistLink="PaulieScanlon/ca0cc9239176066492cb2aba435edbf7" />
+<Tweet tweetLink="PaulieScanlon/status/1201514996838141952" />
+<Instagram instagramId="B7ISOoHne6s" />
+<Pin pinId="637963103444140543" />
